Map/PollutionLayer: 63.0% -> 78.8% (implement wave draw and GX setup)

+++ include/Map/PollutionLayer.hpp
@@ -205,8 +205,8 @@ public:
 	void draw() const;
 	void initGX() const;
 
-	static u32 mInterval;
-	static u32 mAlpha;
+	static f32 mInterval;
+	static u8 mAlpha;
 };
 
 #endif
